Remove local_path.
