+++ tox.ini
@@ -22,6 +22,7 @@ setenv =
 [flake8]
 exclude=migrations,docs
 max-complexity=7
+exclude=migrations
 
 [captainhook]
 pdb=on
